// app/pro-plus/model/job.ts
/**
 * Beacon New Rest Service Ver 1.0
 * Release notes:  `4.8`  * Add `onHold` to request body of endpoint `/submitOrder`
 *
 * OpenAPI spec version: release/4.8
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { Record, String, Boolean, Static } from 'runtypes';

export const Job = Record({
    jobName: String,
    jobNumber: String,
}).Or(
    Record({
        jobName: String,
        jobNumber: String,
        checked: Boolean,
    })
);

export type Job = Static<typeof Job>;

export const noJob: Job = {
    jobName: 'Select Job Account',
    jobNumber: '',
};